move providers to locations in features that make more logical sense
