+++ src/features/toolbar/index.ts
@@ -3,3 +3,4 @@ export * from './Toolbar';
 export * from './RestartPopover';
 export * from './SettingsPopover';
 export * from './SettingSwitch';
+export * from './ShouldRedeal';
