+++ src/features/toolbar/types.ts
@@ -1,7 +1,7 @@
 import { Api as DialogApi } from '@zag-js/dialog';
 import { Api as PopoverApi } from '@zag-js/popover';
-import { Accessor } from 'solid-js';
-import { Setting } from '@/app/types';
+import { Accessor, JSX } from 'solid-js';
+import type { Setting } from '@/features/settings/types';
 
 export interface HowToPlayModalProps {
   machine: Accessor<DialogApi>;
@@ -20,3 +20,15 @@ export interface SettingSwitchProps {
   setting: Setting;
   onSwitch: (arg0: string, arg1: boolean) => void;
 }
+
+export interface RedealProviderProps {
+  shouldRedeal: boolean;
+  children: JSX.Element;
+}
+
+export interface RedealerProvider {
+  shouldRedeal: Accessor<boolean> | (() => boolean);
+  willRedeal: VoidFunction; 
+  willNotRedeal: VoidFunction;
+}
+
